[JS] chore: Begin to move typings to typings folder (#2120)
## Linked issues

#minor  

## Details

We are beginning an effort to better organize the JS package. One step
is moving types to a `/types` folder. This is a partial implementation
of that. This change does not affect users of the Teams AI library.

## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below

---------
