+++ js/packages/teams-ai/src/actions/SayCommand.ts
@@ -6,150 +6,12 @@
  * Licensed under the MIT License.
  */
 
-import { ActivityTypes, Channels, TurnContext } from 'botbuilder';
+import { Activity, ActivityTypes, Channels, TurnContext } from 'botbuilder';
 
 import { PredictedSayCommand } from '../planners';
 import { TurnState } from '../TurnState';
 import { Utilities } from '../Utilities';
-
-export interface AIEntity {
-    /**
-     * Required as 'https://schema.org/Message'
-     */
-    type: 'https://schema.org/Message';
-
-    /**
-     * Required as 'Message
-     */
-    '@type': 'Message';
-
-    /**
-     * Required as 'https://schema.org
-     */
-    '@context': 'https://schema.org';
-
-    /**
-     * Must be left blank. This is for Bot Framework schema.
-     */
-    '@id': '';
-
-    /**
-     * Indicate that the content was generated by AI.
-     */
-    additionalType: ['AIGeneratedContent'];
-
-    /**
-     * Optional; if citations object is included, the  sent activity will include the citations, referenced in the activity text.
-     */
-    citation?: ClientCitation[];
-}
-export interface ClientCitation {
-    /**
-     * Required; must be "Claim"
-     */
-    '@type': 'Claim';
-
-    /**
-     * Required. Number and position of the citation.
-     */
-    position: string;
-    appearance: {
-        /**
-         * Required; Must be 'DigitalDocument'
-         */
-        '@type': 'DigitalDocument';
-
-        /**
-         * Name of the document.
-         */
-        name: string;
-
-        /**
-         * Optional; ignored in Teams
-         */
-        text?: string;
-
-        /**
-         * URL of the document. This will make the name of the citation clickable and direct the user to the specified URL.
-         */
-        url?: string;
-
-        /**
-         * Content of the citation. Must be clipped if longer than 480 characters.
-         */
-        abstract: string;
-
-        /**
-         * Used for icon; for now it is ignored.
-         */
-        encodingFormat?: 'text/html';
-
-        /**
-         * For now ignored, later used for icon
-         */
-        image?: string;
-
-        /**
-         * Optional; set by developer
-         */
-        keywords?: string[];
-
-        /**
-         * Optional sensitivity content information.
-         */
-        usageInfo?: SensitivityUsageInfo;
-    };
-}
-
-/**
- * Sensitivity usage info for content sent to the user. This is used to provide information about the content to the user.
- */
-export interface SensitivityUsageInfo {
-    /**
-     * Must be "https://schema.org/Message"
-     */
-    type: 'https://schema.org/Message';
-
-    /**
-     * Required; Set to CreativeWork;
-     */
-    '@type': 'CreativeWork';
-
-    /**
-     * Sensitivity description of the content
-     */
-    description?: string;
-
-    /**
-     * Sensitivity title of the content
-     */
-    name: string;
-
-    /**
-     * Optional; ignored in Teams.
-     */
-    position?: number;
-
-    pattern?: {
-        /**
-         * Set to DefinedTerm
-         */
-        '@type': 'DefinedTerm';
-
-        inDefinedTermSet: string;
-
-        /**
-         * Color
-         */
-        name: string;
-
-        /**
-         * e.g. #454545
-         */
-        termCode: string;
-    };
-}
-
+import { AIEntity, ClientCitation } from '../types';
 /**
  * @private
  * @param {boolean} feedbackLoopEnabled - If true, the feedback loop UI for Teams will be enabled.
@@ -193,21 +55,24 @@ export function sayCommand<TState extends TurnState = TurnState>(feedbackLoopEna
         // If there are citations, filter out the citations unused in content.
         const referencedCitations = citations ? Utilities.getUsedCitations(contentText, citations) : undefined;
 
-        await context.sendActivity({
+        const entities: AIEntity[] = [
+            {
+                type: 'https://schema.org/Message',
+                '@type': 'Message',
+                '@context': 'https://schema.org',
+                '@id': '',
+                additionalType: ['AIGeneratedContent'],
+                ...(referencedCitations ? { citation: referencedCitations } : {})
+            }
+        ];
+        const activity: Partial<Activity> = {
             type: ActivityTypes.Message,
             text: contentText,
             ...(isTeamsChannel ? { channelData: { feedbackLoopEnabled } } : {}),
-            entities: [
-                {
-                    type: 'https://schema.org/Message',
-                    '@type': 'Message',
-                    '@context': 'https://schema.org',
-                    '@id': '',
-                    additionalType: ['AIGeneratedContent'],
-                    ...(referencedCitations ? { citation: referencedCitations } : {})
-                }
-            ] as AIEntity[]
-        });
+            entities: entities
+        };
+
+        await context.sendActivity(activity);
 
         return '';
     };
