+++ js/packages/teams-ai/src/index.ts
@@ -31,10 +31,4 @@ export * from './TurnState';
 export * from './Utilities';
 export * from './authentication/TeamsBotSsoPrompt';
 export * from './TeamsAdapter';
-export {
-    ActionHandler,
-    ClientCitation,
-    PredictedDoCommandAndHandler,
-    SensitivityUsageInfo,
-    TooManyStepsParameters
-} from './actions';
+export { ActionHandler, PredictedDoCommandAndHandler } from './actions';
