+++ js/packages/teams-ai/src/types/index.ts
@@ -7,5 +7,13 @@
  */
 
 export * from './ActionCall';
+export * from './AIEntity';
+export * from './ClientCitation';
 export * from './CompletionConfig';
+export * from './EmbeddingsModel';
+export * from './InnerMonologue';
+export * from './InnerMonologueSchema';
+export * from './PlanSchema';
 export * from './PromptResponse';
+export * from './SensitivityUsageInfo';
+export * from './TooManyStepsParameters';
